Rename about_me to about.
